Struct hubcaps::git::Git

pub struct Git<'a> { /* fields omitted */ }

reference to git operations associated with a github repo

Methods

impl<'a> Git<'a>

fn tree<S>(&self, sha: S, recursive: bool) -> Result<TreeData> where
S: Into<String>,

list a git tree of files for this repo at a given sha

fn blob<S>(&self, sha: S) -> Result<Blob> where
S: Into<String>,

get the blob contents of a given sha