Module hubcaps::hooks

Hooks interface

See the github docs for more information

Structs

Hook

HookCreateOptions
options for creating a repository hook see this for githubs official documentation

HookCreateOptionsBuilder

HookEditOptions
options for editing a repository hook see this for githubs official documentation

HookEditOptionsBuilder

Hooks
Interface for managing repository hooks

Enums

WebHookContentType
Content-Type web hooks will receive deliveries in