Struct hubcaps::hooks::HookEditOptionsBuilder

pub struct HookEditOptionsBuilder(_);

Methods

impl HookEditOptionsBuilder

pub fn new() -> Self

pub fn active(&mut self, active: bool) -> &mut Self

pub fn events<E>(&mut self, events: Vec<E>) -> &mut Self where
E: Into<String>,

a list of github events this hook should receive deliveries for the default is "push". for a full list, see the Github api docs

pub fn url<U>(&mut self, url: U) -> &mut Self where
U: Into<String>,

web hooks must have an associated url

pub fn content_type(&mut self, content_type: WebHookContentType) -> &mut Self

web hooks can optionally specify a content_type of "form" or "json" which indicates the type of payload they will expect to receive

pub fn secret<S>(&mut self, sec: S) -> &mut Self where
S: Into<String>,

web hooks can optionally provide a secret used to sign deliveries to identify that their source was indeed github

pub fn config_str_entry<K, V>(&mut self, k: K, v: V) -> &mut Self where
K: Into<String>,
V: Into<String>,

pub fn config_entry<N>(&mut self, name: N, value: Value) -> &mut Self where
N: Into<String>,

pub fn build(&self) -> HookEditOptions