Struct hubcaps::issues::IssueRef

pub struct IssueRef<C> where
    C: Clone + Connect + 'static,  { /* fields omitted */ }

Provides access to operations available for a single issue Typically accessed from github.repo(.., ..).issues().get(number)

Methods

impl<C: Clone + Connect + 'static> IssueRef<C>

pub fn get(&self) -> Future<Issue>

Request an issue's information

pub fn labels(&self) -> IssueLabels<C>

Return a reference to labels operations available for this issue

pub fn assignees(&self) -> IssueAssignees<C>

Return a reference to assignee operations available for this issue

pub fn edit(&self, is: &IssueOptions) -> Future<Issue>

Edit the issues options

pub fn comments(&self) -> Comments<C>

Return a reference to comment operations available for this issue