Struct hubcaps::organizations::Organization

pub struct Organization<C> where
    C: Clone + Connect,  { /* fields omitted */ }

Provides access to label operations available for an individual organization

Methods

impl<C: Clone + Connect> Organization<C>

pub fn teams(&self) -> OrgTeams<C>

returns a reference to an interface for team operations

pub fn repos(&self) -> OrgRepositories<C>

returns a reference to an interface for repo operations