Struct hubcaps::pulls::PullRequest

pub struct PullRequest<'a> { /* fields omitted */ }

A structure for accessing interfacing with a specific pull request

Methods

impl<'a> PullRequest<'a>

fn get(&self) -> Result<Pull>

Request a pull requests information

fn open(&self) -> Result<Pull>

short hand for editing state = open

fn close(&self) -> Result<Pull>

shorthand for editing state = closed

fn edit(&self, pr: &PullEditOptions) -> Result<Pull>

Edit a pull request

fn files(&self) -> Result<Vec<FileDiff>>

Returns a vector of file diffs associated with this pull

fn comments(&self) -> Comments

returns issue comments interface

fn review_comments(&self) -> ReviewComments

returns review comments interface

fn commits(&self) -> PullCommits

returns pull commits interface