Struct hubcaps::repositories::Repo

pub struct Repo {
    pub id: u64,
    pub owner: User,
    pub name: String,
    pub full_name: String,
    pub description: Option<String>,
    pub private: bool,
    pub fork: bool,
    pub url: String,
    pub html_url: String,
    pub archive_url: String,
    pub assignees_url: String,
    pub blobs_url: String,
    pub branches_url: String,
    pub clone_url: String,
    pub collaborators_url: String,
    pub comments_url: String,
    pub commits_url: String,
    pub compare_url: String,
    pub contents_url: String,
    pub contributors_url: String,
    pub deployments_url: String,
    pub downloads_url: String,
    pub events_url: String,
    pub forks_url: String,
    pub git_commits_url: String,
    pub git_refs_url: String,
    pub git_tags_url: String,
    pub git_url: String,
    pub hooks_url: String,
    pub issue_comment_url: String,
    pub issue_events_url: String,
    pub issues_url: String,
    pub keys_url: String,
    pub labels_url: String,
    pub languages_url: String,
    pub merges_url: String,
    pub milestones_url: String,
    pub mirror_url: Option<String>,
    pub notifications_url: String,
    pub pulls_url: String,
    pub releases_url: String,
    pub ssh_url: String,
    pub stargazers_url: String,
    pub statuses_url: String,
    pub subscribers_url: String,
    pub subscription_url: String,
    pub svn_url: String,
    pub tags_url: String,
    pub teams_url: String,
    pub trees_url: String,
    pub homepage: Option<String>,
    pub language: Option<String>,
    pub forks_count: u64,
    pub stargazers_count: u64,
    pub watchers_count: u64,
    pub size: u64,
    pub default_branch: String,
    pub open_issues_count: u64,
    pub has_issues: bool,
    pub has_wiki: bool,
    pub has_pages: bool,
    pub has_downloads: bool,
    pub pushed_at: String,
    pub created_at: String,
    pub updated_at: String,
}

Fields
id: u64 owner: User name: String full_name: String description: Option<String> private: bool fork: bool url: String html_url: String archive_url: String assignees_url: String blobs_url: String branches_url: String clone_url: String collaborators_url: String comments_url: String commits_url: String compare_url: String contents_url: String contributors_url: String deployments_url: String downloads_url: String events_url: String forks_url: String git_commits_url: String git_refs_url: String git_tags_url: String git_url: String hooks_url: String issue_comment_url: String issue_events_url: String issues_url: String keys_url: String labels_url: String languages_url: String merges_url: String milestones_url: String mirror_url: Option<String> notifications_url: String pulls_url: String releases_url: String ssh_url: String stargazers_url: String statuses_url: String subscribers_url: String subscription_url: String svn_url: String tags_url: String teams_url: String trees_url: String homepage: Option<String> language: Option<String> forks_count: u64 stargazers_count: u64 watchers_count: u64 size: u64 default_branch: String open_issues_count: u64 has_issues: bool has_wiki: bool has_pages: bool has_downloads: bool pushed_at: String created_at: String updated_at: String
Methods

impl Repo

fn languages(&self, github: &Github) -> Result<HashMap<String, i64>>

Returns a map containing the languages that the repository is implemented in.

The keys are the language names, and the values are the number of bytes of code written in that language.