Struct hubcaps::traffic::Traffic

pub struct Traffic<C> where
    C: Clone + Connect + 'static,  { /* fields omitted */ }

Provides access to the traffic information for a repository

Methods

impl<C: Clone + Connect + 'static> Traffic<C>

pub fn referrers(&self) -> Future<Vec<Referrer>>

List the top 10 referrers over the past 14 days

pub fn paths(&self) -> Future<Vec<Path>>

List the top 10 popular contents over the past 14 days

pub fn views(&self, unit: TimeUnit) -> Future<Views>

Return the total number of views and breakdown per day or week for the last 14 days

pub fn clones(&self, unit: TimeUnit) -> Future<Clones>

Return the total number of clones and breakdown per day or week for the last 14 days