Crate hubcaps

Hubcaps provides a set of building blocks for interacting with the Github API

Reexports

pub use rep::*;

pub use errors::Error;

Modules

deployments
Deployments interface

errors
Client errors

gists
Gists interface

issues
Issues interface

keys
Deploy keys interface This this document for motivation and use

labels
Labels interface

organizations

pulls
Pull requests interface

releases
Releases inteface

rep
Rust representations of Github API data structures

repositories
Repository interface

statuses
Statuses interface

Structs

Github
Entry point interface for interacting with Github API

Enums

Credentials
Various forms of authentication credentials supported by Github

SortDirection
enum representation of Github list sorting options

State
enum representation of github pull and issue state

Type Definitions

Result
alias for Result that infers hubcaps::Error as Err