Struct hubcaps::repositories::Repository

pub struct Repository<'a> {
    // some fields omitted
}

Methods

impl<'a> Repository<'a>

fn new<O, R>(github: &'a Github<'a>, owner: O, repo: R) -> Repository<'a> where O: Into<String>, R: Into<String>

fn deployments(&self) -> Deployments

get a reference to deployments associated with this repository ref

fn issue(&self, number: u64) -> IssueRef

get a reference to a specific github issue associated with this repoistory ref

fn issues(&self) -> Issues

get a reference to github issues associated with this repoistory ref

fn keys(&self) -> Keys

get a reference to deploy keys associated with this repository ref

fn labels(&self) -> Labels

get a list of labels associated with this repository ref

fn pulls(&self) -> PullRequests

get a list of pulls associated with this repository ref

fn releases(&self) -> Releases

get a reference to releases associated with this repository ref

fn statuses(&self) -> Statuses

get a references to statuses associated with this reposoitory ref