Struct hubcaps::repositories::Repository

pub struct Repository { /* fields omitted */ }

Implementations

impl Repository

pub fn get(&self) -> Future<Repo>

get a reference to the GitHub repository object that this Repository refers to

pub fn edit(&self, options: &RepoEditOptions) -> Future<Repo>

https://developer.github.com/v3/repos/#edit

pub fn delete(&self) -> Future<()>

https://developer.github.com/v3/repos/#delete-a-repository

pub fn branches(&self) -> Branches

get a reference to branch operations

pub fn commits(&self) -> RepoCommits

get a reference to repository commit operations

pub fn content(&self) -> Content

get a reference to content operations

pub fn git(&self) -> Git

get a reference to git operations

pub fn hooks(&self) -> Hooks

get a reference to repo hook operations

pub fn deployments(&self) -> Deployments

get a reference to deployments associated with this repository ref

pub fn issue(&self, number: u64) -> IssueRef

get a reference to a specific github issue associated with this repository ref

pub fn issues(&self) -> Issues

get a reference to github issues associated with this repository ref

pub fn checkruns(&self) -> CheckRuns

get a reference to github checks associated with this repository ref

pub fn keys(&self) -> Keys

get a reference to deploy keys associated with this repository ref

pub fn labels(&self) -> Labels

get a list of labels associated with this repository ref

pub fn pulls(&self) -> PullRequests

get a list of pulls associated with this repository ref

pub fn releases(&self) -> Releases

get a reference to releases associated with this repository ref

pub fn statuses(&self) -> Statuses

get a reference to statuses associated with this repository ref

pub fn collaborators(&self) -> Collaborators

get a reference to the collaborators associated with this repository ref

pub fn teams(&self) -> RepoTeams

get a reference to teams associated with this repository ref

pub fn contributors(&self) -> Contributors

get a reference to contributors associated with this repository ref

pub fn contributor_statistics(&self) -> ContributorStatistics

pub fn traffic(&self) -> Traffic

get a reference of traffic associated with this repository ref

pub fn forks(&self) -> Forks

get a reference to forks associated with this repository ref