Struct hubcaps::Github

pub struct Github<'a> {
    // some fields omitted
}

Entry point interface for interacting with Github API

Methods

impl<'a> Github<'a>

fn new<A>(agent: A, client: &'a Client, credentials: Credentials) -> Github<'a> where A: Into<String>

Create a new Github instance

fn host<H, A>(host: H, agent: A, client: &'a Client, credentials: Credentials) -> Github<'a> where H: Into<String>, A: Into<String>

Create a new Github instance hosted at a custom location. Useful for github enterprise installations ( yourdomain.com/api/v3/ )

fn repo<O, R>(&self, owner: O, repo: R) -> Repository where O: Into<String>, R: Into<String>

Return a reference to a Github reposistory

fn user_repos<S>(&self, owner: S) -> UserRepositories where S: Into<String>

Return a reference to the collection of repositories owned by and associated with an owner

fn repos(&self) -> Repositories

Return a reference to the collection of repositories owned by the user associated with the current authentication credentials

fn orgs(&self) -> Organizations

Return a reference to the collection of organizations that the user associated with the current authentication credentials is in

fn user_orgs<U>(&self, user: U) -> UserOrganizations where U: Into<String>

Return a reference to the collection of organizations a user is publicly associated with

fn user_gists<O>(&self, owner: O) -> UserGists where O: Into<String>

Return a reference to an interface that provides access to a user's gists

fn gists(&self) -> Gists

Return a reference to an interface that provides access to the gists belonging to the owner of the token used to configure this client

fn org_repos<O>(&self, org: O) -> OrganizationRepositories where O: Into<String>

Return a reference to the collection of repositories owned by and associated with an organization