Struct hubcaps::Github

pub struct Github { /* fields omitted */ }

Entry point interface for interacting with GitHub API

Implementations

impl Github

pub fn new<A, C>(agent: A, credentials: C) -> Result<Self> where
A: Into<String>,
C: Into<Option<Credentials>>,

pub fn host<H, A, C>(host: H, agent: A, credentials: C) -> Result<Self> where
H: Into<String>,
A: Into<String>,
C: Into<Option<Credentials>>,

pub fn custom<H, A, CR>(
host: H,
agent: A,
credentials: CR,
http: Client
) -> Self where
H: Into<String>,
A: Into<String>,
CR: Into<Option<Credentials>>,

pub fn set_credentials<CR>(&mut self, credentials: CR) where
CR: Into<Option<Credentials>>,

pub fn rate_limit(&self) -> RateLimit

pub fn activity(&self) -> Activity

Return a reference to user activity

pub fn repo<O, R>(&self, owner: O, repo: R) -> Repository where
O: Into<String>,
R: Into<String>,

Return a reference to a Github repository

pub fn user_repos<S>(&self, owner: S) -> UserRepositories where
S: Into<String>,

Return a reference to the collection of repositories owned by and associated with an owner

pub fn repos(&self) -> Repositories

Return a reference to the collection of repositories owned by the user associated with the current authentication credentials

pub fn org<O>(&self, org: O) -> Organization where
O: Into<String>,

pub fn orgs(&self) -> Organizations

Return a reference to the collection of organizations that the user associated with the current authentication credentials is in

pub fn users(&self) -> Users

Return a reference to an interface that provides access to user information.

pub fn user_orgs<U>(&self, user: U) -> UserOrganizations where
U: Into<String>,

Return a reference to the collection of organizations a user is publicly associated with

pub fn user_gists<O>(&self, owner: O) -> UserGists where
O: Into<String>,

Return a reference to an interface that provides access to a user's gists

pub fn gists(&self) -> Gists

Return a reference to an interface that provides access to the gists belonging to the owner of the token used to configure this client

pub fn search(&self) -> Search

Return a reference to an interface that provides access to search operations

pub fn org_repos<O>(&self, org: O) -> OrganizationRepositories where
O: Into<String>,

Return a reference to the collection of repositories owned by and associated with an organization

pub fn app(&self) -> App

Return a reference to GitHub Apps