Struct hubcaps::Github

pub struct Github<C> where
    C: Clone + Connect + 'static,  { /* fields omitted */ }

Entry point interface for interacting with Github API

Methods

impl Github<HttpsConnector<HttpConnector>>

pub fn new<A, C>(agent: A, credentials: C) -> Self where
A: Into<String>,
C: Into<Option<Credentials>>,

pub fn host<H, A, C>(host: H, agent: A, credentials: C) -> Self where
H: Into<String>,
A: Into<String>,
C: Into<Option<Credentials>>,

impl<C> Github<C> where
C: Clone + Connect + 'static,

pub fn custom<H, A, CR>(
host: H,
agent: A,
credentials: CR,
http: Client<C>
) -> Self where
H: Into<String>,
A: Into<String>,
CR: Into<Option<Credentials>>,

pub fn rate_limit(&self) -> RateLimit<C>

pub fn activity(&self) -> Activity<C>

Return a reference to user activity

pub fn repo<O, R>(&self, owner: O, repo: R) -> Repository<C> where
O: Into<String>,
R: Into<String>,

Return a reference to a Github repository

pub fn user_repos<S>(&self, owner: S) -> UserRepositories<C> where
S: Into<String>,

Return a reference to the collection of repositories owned by and associated with an owner

pub fn repos(&self) -> Repositories<C>

Return a reference to the collection of repositories owned by the user associated with the current authentication credentials

pub fn org<O>(&self, org: O) -> Organization<C> where
O: Into<String>,

pub fn orgs(&self) -> Organizations<C>

Return a reference to the collection of organizations that the user associated with the current authentication credentials is in

pub fn users(&self) -> Users<C>

Return a reference to an interface that provides access to user information.

pub fn user_orgs<U>(&self, user: U) -> UserOrganizations<C> where
U: Into<String>,

Return a reference to the collection of organizations a user is publicly associated with

pub fn user_gists<O>(&self, owner: O) -> UserGists<C> where
O: Into<String>,

Return a reference to an interface that provides access to a user's gists

pub fn gists(&self) -> Gists<C>

Return a reference to an interface that provides access to the gists belonging to the owner of the token used to configure this client

pub fn search(&self) -> Search<C>

Return a reference to an interface that provides access to search operations

pub fn org_repos<O>(&self, org: O) -> OrganizationRepositories<C> where
O: Into<String>,

Return a reference to the collection of repositories owned by and associated with an organization